+++ oedatamodel_api/upload.py
@@ -109,6 +109,8 @@ def upload_data_to_oep(data, schema):
             return item.isoformat()
         return item
 
+    # TODO: Check for type "Decimal" and report error; solution in datapacke.json: "floatNumber": "True"
+
     for table, table_data in data.items():
         table_url = f"{OEP_URL}/api/v0/schema/{schema}/tables/{table}/rows/new"
         response = requests.post(